ai8x-training/requirements.txt
# torch==2.3.1
# torchvision==0.18.1
torchaudio==2.3.1
GitPython==3.1.43
Pillow==10.4.0
PyYAML==6.0.1
albumentations==1.4.15
albucore==0.0.16
faiss-cpu==1.8.0.post1
batch-face==1.4.0
h5py==3.11.0
kornia==0.7.3
numpy==1.26.4
opencv-python==4.10.0.84
pycocotools==2.0.8
# pyffmpeg==2.4.2.18.1
pytorch-metric-learning==2.6.0
pytube==15.0.0
qrcode==7.4.2
seaborn==0.13.2
scipy==1.14.0
soundfile==0.12.1
tensorboard==2.17.0
tk==0.1.0
torchmetrics==1.4.0.post0
torchnet==0.0.4
-e distiller --config-settings editable_mode=strict
